to implement TP

+++ workspace/mistral/engine/transformer_layers.py
@@ -123,7 +123,7 @@ class RMSNorm(torch.nn.Module):
         return x * torch.rsqrt(x.pow(2).mean(-1, keepdim=True) + self.eps)
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
-        output = self._norm(x.float()).type_as(x)
+        output = self._norm(x.float()).to(x.dtype)
         return output * self.weight
